Move files into their final location
